+++ examples/react/src/pages/active-wallet.tsx
@@ -1,17 +1,19 @@
 import { useChain, useWalletModal } from "@interchain-kit/react";
 
 export default function ActiveWallet() {
-  const { address, wallet, logoUrl } = useChain("juno");
+  const { address, wallet, logoUrl, connect, disconnect } = useChain("juno");
   const { open } = useWalletModal();
   return (
     <>
       <button onClick={open}>open wallets modal</button>
       <p>{wallet?.walletState}</p>
-      <p>{wallet?.option?.prettyName}</p>
+      <p>{wallet?.info?.prettyName}</p>
 
       <p>{address}</p>
-      <img src={wallet?.option?.logo?.toString()} style={{ width: "100px" }} />
+      <img src={wallet?.info?.logo?.toString()} style={{ width: "100px" }} />
       <img src={logoUrl} style={{ width: "100px" }}></img>
+      <button onClick={connect}>Connect</button>
+      <button onClick={disconnect}>Disconnect</button>
     </>
   );
 }