+++ packages/react/src/modal/views/WalletList.tsx
@@ -4,7 +4,7 @@ import { useWalletManager } from "../../hooks";
 import { BaseWallet } from "@interchain-kit/core";
 
 export const WalletListHeader = () => {
-  const { close } = useWalletModal()
+  const { close } = useWalletModal();
   return (
     <ConnectModalHead
       title="Select your wallet"
@@ -12,33 +12,37 @@ export const WalletListHeader = () => {
       onClose={close}
       closeButtonProps={{ onClick: close }}
     />
-  )
-}
+  );
+};
 
 export const WalletListContent = () => {
-
-  const walletManager = useWalletManager()
-  const { close } = useWalletModal()
+  const walletManager = useWalletManager();
+  const { close } = useWalletModal();
 
   const wallets = walletManager.wallets.map((w: BaseWallet) => {
-    return ({
-      name: w.option.name,
-      prettyName: w.option.prettyName,
-      logo: w.option.logo as string,
+    return {
+      name: w.info.name,
+      prettyName: w.info.prettyName,
+      logo: w.info.logo as string,
       mobileDisabled: true,
-      shape: 'list' as 'list',
-      originalWallet: w
-    })
-  })
+      shape: "list" as "list",
+      originalWallet: w,
+    };
+  });
 
   const onWalletClick = async (wallet: any) => {
     try {
-      await walletManager.connect(wallet.option.name)
-      close()
+      await walletManager.connect(wallet.info.name);
+      close();
     } catch (error) {
-      console.log(error)
+      console.log(error);
     }
-  }
+  };
 
-  return <ConnectModalWalletList wallets={wallets} onWalletItemClick={onWalletClick} />
-}
+  return (
+    <ConnectModalWalletList
+      wallets={wallets}
+      onWalletItemClick={onWalletClick}
+    />
+  );
+};