+++ packages/react/src/types/sign-client.ts
@@ -1,0 +1,3 @@
+import { WalletManager } from "@interchain-kit/core";
+
+export type SigningClient = Awaited<ReturnType<WalletManager['getSigningClient']>>