+++ packages/react/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.1-beta.33](https://github.com/interchain-kit/react/compare/@interchain-kit/react@0.0.1-beta.32...@interchain-kit/react@0.0.1-beta.33) (2024-10-21)
+
+**Note:** Version bump only for package @interchain-kit/react
+
+## [0.0.1-beta.32](https://github.com/interchain-kit/react/compare/@interchain-kit/react@0.0.1-beta.31...@interchain-kit/react@0.0.1-beta.32) (2024-10-21)
+
+**Note:** Version bump only for package @interchain-kit/react
+
+## [0.0.1-beta.31](https://github.com/interchain-kit/react/compare/@interchain-kit/react@0.0.1-beta.30...@interchain-kit/react@0.0.1-beta.31) (2024-10-21)
+
+**Note:** Version bump only for package @interchain-kit/react
+
+## [0.0.1-beta.30](https://github.com/interchain-kit/react/compare/@interchain-kit/react@0.0.1-beta.29...@interchain-kit/react@0.0.1-beta.30) (2024-10-21)
+
+**Note:** Version bump only for package @interchain-kit/react
+
 ## [0.0.1-beta.29](https://github.com/interchain-kit/react/compare/@interchain-kit/react@0.0.1-beta.28...@interchain-kit/react@0.0.1-beta.29) (2024-10-18)
 
 **Note:** Version bump only for package @interchain-kit/react